+++ src/app/App.tsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { Route, Routes } from 'react-router-dom';
+import AboutUs from '../components/pages/about/AboutUs';
+import FormsPage from '../components/pages/form-page/FormsPage';
+import Main from '../components/pages/main/Main';
+import NotFound from '../components/pages/notFound/NotFound';
 import './App.scss';
-import Main from '../pages/main/Main';
-import NotFound from '../pages/notFound/NotFound';
-import AboutUs from '../pages/about/AboutUs';
-import Forms from '../pages/forms/Forms';
 
 export default class App extends Component {
   render() {
@@ -12,7 +12,7 @@ export default class App extends Component {
       <Routes>
         <Route path="/" element={<Main />} />
         <Route path="/about-us" element={<AboutUs />} />
-        <Route path="/forms" element={<Forms />} />
+        <Route path="/forms" element={<FormsPage />} />
         <Route path="*" element={<NotFound />} />
       </Routes>
     );
