+++ src/common/hoc/withHeaderHoc.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Header from '../header/Header';
+import Header from '../../components/header/Header';
 
 export const withHeaderHoc = (ViewComponent: React.ComponentType) => {
   return class extends Component {
