+++ src/components/card-list/CardList.tsx
@@ -6,17 +6,29 @@ import './card-list.scss';
 export default class CardList extends Component<CardProps> {
   render() {
     const { data, searchValue } = this.props;
-
+    const filterField = [
+      'name',
+      'collection',
+      'category',
+      'metal',
+      'rating',
+      'year',
+      'price',
+    ];
     const filterData = searchValue.trim()
       ? data.filter((el) => {
-          const values = Object.values(el).map((e) =>
-            e.toString().toUpperCase()
-          );
-          return values.some((item) =>
-            item.includes(searchValue.toUpperCase())
+          const values = Object.entries(el).map((e) => [
+            e[0],
+            e[1].toString().toUpperCase(),
+          ]);
+          return values.some(
+            (item) =>
+              filterField.includes(item[0]) &&
+              item[1].includes(searchValue.toUpperCase())
           );
         })
       : data;
+
     return (
       <ul className="cardsList">
         {filterData.map((el) => {
