+++ src/components/pages/form-page/FormsPage.tsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import { withHeaderHoc } from '../../../common/hoc/withHeaderHoc';
 
-class AboutUs extends Component {
+class FormsPage extends Component {
   render() {
     return (
       <>
-        <main>AboutUs</main>
+        <main>Forms</main>
       </>
     );
   }
 }
-export default withHeaderHoc(AboutUs);
+export default withHeaderHoc(FormsPage);
