--- src/components/pages/forms/Forms.tsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-import { withHeaderHoc } from '../../hoc/withHeaderHoc';
-
-class Forms extends Component {
-  render() {
-    return (
-      <>
-        <main>Forms</main>
-      </>
-    );
-  }
-}
-export default withHeaderHoc(Forms);
