+++ src/components/searchForn/searchForm.scss
@@ -7,11 +7,7 @@
         color:#ffffff;
         font-size: 1.5rem;
         margin-right: 10px;
-    }
-    &__button{
-        height: 2rem;
-        background-color: antiquewhite;
-        font-size: 1.5rem;
-        font-weight: bold;
+        position: relative;
+        top: -2rem;
     }
 }
