+++ tsconfig.json
@@ -12,7 +12,7 @@
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true,
-    // "baseUrl": "./src",
+    "sourceMap": true,
     "resolveJsonModule": true
   },
   "include": ["vite.config.ts", ".eslintrc.cjs", "src"]
